+++ docs/notebooks/iris-rfc/config.yaml
@@ -1,0 +1,42 @@
+bundled_packages_dir: packages
+data_dir: data
+description: null
+environment_variables: {}
+examples_filename: examples.yaml
+external_package_dirs: []
+input_type: Any
+live_reload: false
+model_class_filename: model.py
+model_class_name: Model
+model_framework: sklearn
+model_metadata:
+  model_binary_dir: model
+  supports_predict_proba: true
+model_module_dir: model
+model_name: null
+model_type: Model
+python_version: py39
+requirements:
+- scikit-learn==1.0.2
+- threadpoolctl==3.1.0
+- joblib==1.2.0
+- numpy==1.23.5
+- scipy==1.10.0
+resources:
+  accelerator: null
+  cpu: 500m
+  memory: 512Mi
+  use_gpu: false
+secrets: {}
+spec_version: '2.0'
+system_packages: []
+train:
+  resources:
+    accelerator: null
+    cpu: 500m
+    memory: 512Mi
+    use_gpu: false
+  training_class_filename: train.py
+  training_class_name: Train
+  training_module_dir: train
+  variables: {}
