+++ truss/templates/keras/model/model.py
@@ -1,4 +1,4 @@
-from typing import Dict, List
+from typing import Any, Dict, List
 
 import numpy as np
 from tensorflow import keras
@@ -17,24 +17,23 @@ class Model:
             str(self._data_dir / self._model_binary_dir)
         )
 
-    def preprocess(self, request: Dict) -> Dict:
+    def preprocess(self, model_input: Any) -> Any:
         """
         Incorporate pre-processing required by the model if desired here.
 
         These might be feature transformations that are tightly coupled to the model.
         """
-        return request
+        return model_input
 
-    def postprocess(self, request: Dict) -> Dict:
+    def postprocess(self, model_output: Dict) -> Dict:
         """
         Incorporate post-processing required by the model if desired here.
         """
-        return request
+        return model_output
 
-    def predict(self, request: Dict) -> Dict[str, List]:
-        response = {}
-        inputs = request["inputs"]
-        inputs = np.array(inputs)
+    def predict(self, model_input: Any) -> Dict[str, List]:
+        model_output = {}
+        inputs = np.array(model_input)
         result = self._model.predict(inputs).tolist()
-        response["predictions"] = result
-        return response
+        model_output["predictions"] = result
+        return model_output
