+++ truss/templates/lightgbm/model/model.py
@@ -1,4 +1,4 @@
-from typing import Dict, List
+from typing import Any, Dict, List
 
 MODEL_BASENAME = "model"
 MODEL_EXTENSIONS = [".joblib", ".pkl", ".pickle"]
@@ -25,25 +25,26 @@ class Model:
         model_file_path = next(path for path in paths if path.exists())
         self._model = joblib.load(model_file_path)
 
-    def preprocess(self, request: Dict) -> Dict:
+    def preprocess(self, model_input: Any) -> Any:
         """
         Incorporate pre-processing required by the model if desired here.
 
         These might be feature transformations that are tightly coupled to the model.
         """
-        return request
+        return model_input
 
-    def postprocess(self, request: Dict) -> Dict:
+    def postprocess(self, model_output: Dict) -> Dict:
         """
         Incorporate post-processing required by the model if desired here.
         """
-        return request
+        return model_output
 
-    def predict(self, request: Dict) -> Dict[str, List]:
-        response = {}
-        inputs = request["inputs"]
-        result = self._model.predict(inputs)
-        response["predictions"] = result
+    def predict(self, model_input: Any) -> Dict[str, List]:
+        model_output = {}
+        result = self._model.predict(model_input)
+        model_output["predictions"] = result
         if self._supports_predict_proba:
-            response["probabilities"] = self._model.predict_proba(inputs).tolist()
-        return response
+            model_output["probabilities"] = self._model.predict_proba(
+                model_input
+            ).tolist()
+        return model_output
