+++ truss/templates/mlflow/model/model.py
@@ -1,4 +1,4 @@
-from typing import Dict, List
+from typing import Any, Dict, List
 
 import mlflow
 import numpy as np
@@ -16,24 +16,23 @@ class Model:
         model_binary_dir_path = self._data_dir / self._model_binary_dir
         self._model = mlflow.pyfunc.load_model(model_binary_dir_path / "model")
 
-    def preprocess(self, request: Dict) -> Dict:
+    def preprocess(self, model_input: Any) -> Any:
         """
         Incorporate pre-processing required by the model if desired here.
 
         These might be feature transformations that are tightly coupled to the model.
         """
-        return request
+        return model_input
 
-    def postprocess(self, request: Dict) -> Dict:
+    def postprocess(self, model_output: Dict) -> Dict:
         """
         Incorporate post-processing required by the model if desired here.
         """
-        return request
+        return model_output
 
-    def predict(self, request: Dict) -> Dict[str, List]:
-        response = {}
-        inputs = request["inputs"]
-        inputs = np.array(inputs)
+    def predict(self, model_input: Any) -> Dict[str, List]:
+        model_output = {}
+        inputs = np.array(model_input)
         result = self._model.predict(inputs).tolist()
-        response["predictions"] = result
-        return response
+        model_output["predictions"] = result
+        return model_output
