+++ truss/templates/pytorch/model/model.py
@@ -1,4 +1,4 @@
-from typing import Dict, List
+from typing import Any, Dict, List
 
 import torch
 from torch import package
@@ -26,26 +26,25 @@ class Model:
         self._model = imp.load_pickle(package_name, model_pickle_filename)
         self._model_dtype = list(self._model.parameters())[0].dtype
 
-    def preprocess(self, request: Dict) -> Dict:
+    def preprocess(self, model_input: Any) -> Any:
         """
         Incorporate pre-processing required by the model if desired here.
 
         These might be feature transformations that are tightly coupled to the model.
         """
-        return request
+        return model_input
 
-    def postprocess(self, request: Dict) -> Dict:
+    def postprocess(self, model_output: Dict) -> Dict:
         """
         Incorporate post-processing required by the model if desired here.
         """
-        return request
+        return model_output
 
-    def predict(self, request: Dict) -> Dict[str, List]:
-        response = {}
+    def predict(self, model_input: Any) -> Dict[str, List]:
+        model_output = {}
         with torch.no_grad():
             inputs = torch.tensor(
-                request["inputs"], dtype=self._model_dtype, device=self._device
+                model_input, dtype=self._model_dtype, device=self._device
             )
-            response = {}
-            response["predictions"] = self._model(inputs).tolist()
-            return response
+            model_output["predictions"] = self._model(inputs).tolist()
+            return model_output
