+++ truss/templates/xgboost/model/model.py
@@ -1,4 +1,4 @@
-from typing import Dict, List
+from typing import Any, Dict, List
 
 import xgboost as xgb
 
@@ -31,24 +31,23 @@ class Model:
         self._model = xgb.Booster()
         self._model.load_model(model_file_path)
 
-    def preprocess(self, request: Dict) -> Dict:
+    def preprocess(self, model_input: Any) -> Any:
         """
         Incorporate pre-processing required by the model if desired here.
 
         These might be feature transformations that are tightly coupled to the model.
         """
-        return request
+        return model_input
 
-    def postprocess(self, request: Dict) -> Dict:
+    def postprocess(self, model_output: Dict) -> Dict:
         """
         Incorporate post-processing required by the model if desired here.
         """
-        return request
+        return model_output
 
-    def predict(self, request: Dict) -> Dict[str, List]:
-        response = {}
-        inputs = request["inputs"]
-        dmatrix_inputs = xgb.DMatrix(inputs)
+    def predict(self, model_input: Any) -> Dict[str, List]:
+        model_output = {}
+        dmatrix_inputs = xgb.DMatrix(model_input)
         result = self._model.predict(dmatrix_inputs)
-        response["predictions"] = result
-        return response
+        model_output["predictions"] = result
+        return model_output
